+++ thirdeye-bootstrap/src/main/java/com/linkedin/thirdeye/bootstrap/rollup/AverageBasedRollupFunction.java
@@ -23,7 +23,7 @@ public class AverageBasedRollupFunction implements RollupThresholdFunc {
     Set<Long> timeWindowSet = timeSeries.getTimeWindowSet();
     long sum = 0;
     for (Long timeWindow : timeWindowSet) {
-      sum += timeSeries.get(timeWindow, metricName);
+      sum += timeSeries.get(timeWindow, metricName).longValue();
     }
     return sum / timeWindowSet.size() >= averageThreshold;
   }
